refactor(weather): split aviation enrichment helpers
Move aviation TAF routing, advisory filtering, and AVWX/NWS fallback logic into a focused module while keeping weather_client_enrichment compatibility exports stable.

Tested: focused aviation/enrichment/weather-client pytest slice; ruff check src tests scripts; pyright; broad tests excluding known native toast/audio/thread crash cluster.

Confidence: high
